requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
Django==3.1.1
idna==2.10
gunicorn
multidict==4.7.6
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
yarl==1.5.1
whitenoise
pymongo
dnspython